Address web.py estimate gas failings post 1559
I am a noob to this process so I am sure the code could be improved to more precisely get the "maxFeePerGas" and "maxPriorityFeePerGas" and "gas" however the current web3.py library throws an error when estimating gas in the build_transaction function. With these edits, the demo code at least works.